git-svn-id: svn://svn.hafro.is/jminor/branches/java11@17517 51816742-9f83-410d-85ed-e7bcf7042953

+++ swing/common-ui/src/main/java/org/jminor/swing/common/ui/input/DoubleInputProvider.java
@@ -3,12 +3,12 @@
  */
 package org.jminor.swing.common.ui.input;
 
-import org.jminor.swing.common.ui.textfield.DoubleField;
+import org.jminor.swing.common.ui.textfield.DecimalField;
 
 /**
  * A InputProvider implementation for double values.
  */
-public final class DoubleInputProvider extends AbstractInputProvider<Double, DoubleField> {
+public final class DoubleInputProvider extends AbstractInputProvider<Double, DecimalField> {
 
   /**
    * Instantiates a new DoubleInputProvider.
@@ -25,7 +25,7 @@ public final class DoubleInputProvider extends AbstractInputProvider<Double, Dou
    * @param maxValue the maximum value
    */
   public DoubleInputProvider(final Double initialValue, final double minValue, final double maxValue) {
-    super(new DoubleField());
+    super(new DecimalField());
     getInputComponent().setRange(minValue, maxValue);
     getInputComponent().setDouble(initialValue);
   }
